Revert "Switched from C to C++"
This reverts commit 947788a252578747837dbc8b8c28b497f920dd85.

I switched this project to C++ with the goal of expanding the demos with
code which would do more complicated encoding and decoding of CAN
messages. I never followed through with that and ended up leaving this
project in a bad state where some demos failed to build.

Note, the demos failed to build using a C++ compiler because the
`bcm_msg_head` struct uses a flexible array member which can't be
allocated the same way as it's done in C.

Now, I'm reverting this project back to C and working on just enhancing
the demos I had previously written.